refactor: 기타 파일 컨벤션 맞춤
