+++ src/common/shared-kernel/interfaces/use-case.response.ts
