+++ src/common/shared-kernel/interfaces/use-case.ts
