+++ src/common/system/persistences/entities/counselors/tone.entity.ts
@@ -1,8 +1,8 @@
 import { CoreEntity } from "~common/system/persistences/entities/core.entity";
 import { CounselorEntity } from "~common/system/persistences/entities/counselors/counselor.entity";
 import { CounselTechniquesEntity } from "~common/system/persistences/entities/prompts/counsel-techniques.entity";
-import { TonePromptEntity } from "~common/system/persistences/entities/prompts/TonePrompts.entity";
-import { ToneScopedPromptEntity } from "~common/system/persistences/entities/prompts/ToneScopedPrompts.entity";
+import { TonePromptEntity } from "~common/system/persistences/entities/prompts/tone-prompts.entity";
+import { ToneScopedPromptEntity } from "~common/system/persistences/entities/prompts/tone-scoped-prompts.entity";
 import { Column, Entity, OneToMany } from "typeorm";
 
 @Entity({
