+++ src/services/counselings/applications/counsel-prompt-managements/use-cases/dtos/validate-prompt-version.dto.ts
@@ -1,6 +1,6 @@
 import { PromptVersionInfo } from "~counselings/domains/prompt-versions/models/prompt-version.info";
 
-import { UseCaseCoreResponse } from "~common/shared-kernel/interfaces/UseCase.response";
+import { UseCaseCoreResponse } from "~common/shared-kernel/interfaces/use-case.response";
 
 export interface ValidatePromptVersionRequest {
   promptVersion: PromptVersionInfo;
