+++ src/services/counselings/applications/counsel-prompt-managements/use-cases/validate-prompt-version.ts
@@ -6,7 +6,7 @@ import { CounselorsService } from "~counselings/domains/counselors/counselors.se
 import { TonesService } from "~counselings/domains/tones/tones.service";
 
 import { HttpStatus, Injectable } from "@nestjs/common";
-import { UseCase } from "~common/shared-kernel/interfaces/UseCase";
+import { UseCase } from "~common/shared-kernel/interfaces/use-case";
 import { HttpStatusBasedRpcException } from "~common/system/filters/exceptions";
 
 @Injectable()
