+++ src/services/counselings/counsels.service.module.ts
@@ -1,12 +1,12 @@
 import { CounselManagementsModule } from "~counselings/applications/counsel-managements/counsel-managements.module";
 import { CounselPromptManagementsModule } from "~counselings/applications/counsel-prompt-managements/counsel-prompt-managements.module";
 import { CounselorManagementsModule } from "~counselings/applications/counselor-managements/counselor-managements.module";
+import { GrpcCounselPromptCommandController } from "~counselings/presentations/grpc/command/counsel-prompts-command.controller";
 import { GrpcCounselorCommandController } from "~counselings/presentations/grpc/command/counselors-command.controller";
-import { GrpcCounselPromptCommandController } from "~counselings/presentations/grpc/command/counselPrompts-command.controller";
 import { GrpcCounselCommandController } from "~counselings/presentations/grpc/command/counsels-command.controller";
 import { GrpcEpisodeCommandController } from "~counselings/presentations/grpc/command/episodes-command.controller";
+import { GrpcCounselPromptQueryController } from "~counselings/presentations/grpc/query/counsel-prompts-query.controller";
 import { GrpcCounselorQueryController } from "~counselings/presentations/grpc/query/counselors-query.controller";
-import { GrpcCounselPromptQueryController } from "~counselings/presentations/grpc/query/counselPrompts-query.controller";
 import { GrpcCounselQueryController } from "~counselings/presentations/grpc/query/counsels-query.controller";
 import { GrpcEpisodeQueryController } from "~counselings/presentations/grpc/query/episodes-query.controller";
