+++ src/services/counselings/domains/counsels/infrastructures/counsels.repository.ts
@@ -2,7 +2,7 @@ import { Counsels } from "~counselings/domains/counsels/models/counsels";
 
 import { Injectable } from "@nestjs/common";
 import { CounselId } from "~common/shared-kernel/identifiers/counsel.id";
-import { CounselsEntity } from "~common/system/persistences/entities/counsels/Counsels.entity";
+import { CounselsEntity } from "~common/system/persistences/entities/counsels/counsel.entity";
 import { FindManyOptions, FindOneOptions } from "typeorm";
 
 @Injectable()
