+++ src/services/counselings/domains/persona-prompts/infrastructures/persona-prompts.repository.ts
@@ -2,7 +2,7 @@ import { PersonaPrompts } from "~counselings/domains/persona-prompts/models/pers
 
 import { Injectable } from "@nestjs/common";
 import { PersonaPromptId } from "~common/shared-kernel/identifiers/persona-prompt.id";
-import { PersonaPromptEntity } from "~common/system/persistences/entities/prompts/PersonaPrompts.entity";
+import { PersonaPromptEntity } from "~common/system/persistences/entities/prompts/persona-prompts.entity";
 import { FindManyOptions, FindOneOptions } from "typeorm";
 
 @Injectable()
