+++ src/services/counselings/domains/prompt-activate-history/infrastructures/typeorm-prompt-activate-history.repository.ts
@@ -5,7 +5,7 @@ import { PromptActivateHistories } from "~counselings/domains/prompt-activate-hi
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { PromptActivateHistoryId } from "~common/shared-kernel/identifiers/prompt-activate-history.id";
-import { PromptActivateHistoryEntity } from "~common/system/persistences/entities/prompts/PromptActivateHistory.entity";
+import { PromptActivateHistoryEntity } from "~common/system/persistences/entities/prompts/prompt-activate-history.entity";
 import { FindOneOptions, Repository } from "typeorm";
 
 @Injectable()
