+++ src/services/counselings/domains/tone-prompts/infrastructures/mappers/repository-tonePrompts-criteria.mapper.ts
@@ -1,6 +1,6 @@
 import { TonePromptsCriteriaFindMany } from "~counselings/domains/tone-prompts/tone-prompts.criteria";
 
-import { TonePromptEntity } from "~common/system/persistences/entities/prompts/TonePrompts.entity";
+import { TonePromptEntity } from "~common/system/persistences/entities/prompts/tone-prompts.entity";
 import { FindManyOptions, FindOptionsWhere } from "typeorm";
 
 export class RepositoryTonePromptCriteriaMapper {
