+++ src/services/counselings/domains/tone-prompts/infrastructures/tone-prompts.repository.ts
@@ -2,7 +2,7 @@ import { TonePrompts } from "~counselings/domains/tone-prompts/models/tone-promp
 
 import { Injectable } from "@nestjs/common";
 import { TonePromptId } from "~common/shared-kernel/identifiers/tone-prompt.id";
-import { TonePromptEntity } from "~common/system/persistences/entities/prompts/TonePrompts.entity";
+import { TonePromptEntity } from "~common/system/persistences/entities/prompts/tone-prompts.entity";
 import { FindManyOptions, FindOneOptions } from "typeorm";
 
 @Injectable()
